+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- A skinnable bar during gameplay to represent the time left in the map.
+
 ### Changed
 - Scrolling in the song select or settings menu is no longer infinite.